% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_adminModal.R
\name{adminModal}
\alias{adminModal}
\title{Admin Management Modal UI}
\usage{
adminModal(id)
}
\arguments{
\item{id}{\code{character} identifier for the modal}
}
\value{
A \code{shiny.tag} object containing the admin modal UI
}
\description{
Creates the admin management modal UI for managers
}
