% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_adminModal.R
\name{adminModalServer}
\alias{adminModalServer}
\title{Admin Management Modal Server}
\usage{
adminModalServer(id, client, auth_token, user_info)
}
\arguments{
\item{id}{\code{character} identifier for the module}

\item{client}{\code{GraphqlClient} object for GraphQL communication}

\item{auth_token}{\code{reactiveVal} to store authentication token}

\item{user_info}{\code{reactiveVal} to store user information}
}
\description{
Handles the admin management modal server logic
}
