% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_adminPage.R
\name{adminPageUI}
\alias{adminPageUI}
\title{Admin Page UI}
\usage{
adminPageUI(id)
}
\arguments{
\item{id}{\code{character} identifier.}
}
\value{
A \code{shiny.tag} object.
}
\description{
Creates the admin page interface for managing projects
}
