% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_aoiSelection.R
\name{aoiSelectionServer}
\alias{aoiSelectionServer}
\title{AOI Selection Server}
\usage{
aoiSelectionServer(id, app_data, session, map_id = "map")
}
\arguments{
\item{id}{\code{character} identifier.}

\item{app_data}{\code{environment} containing application data.}

\item{session}{\code{ShinySession} object.}

\item{map_id}{\code{character} identifier for the map.}
}
\description{
Server logic for Area of Interest (AOI) selection and analysis
}
