% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_aoiSelection.R
\name{aoiSelectionUI}
\alias{aoiSelectionUI}
\title{AOI Selection UI}
\usage{
aoiSelectionUI(id)
}
\arguments{
\item{id}{\code{character} identifier.}
}
\value{
A \code{shiny.tag} object.
}
\description{
Creates the Area of Interest (AOI) selection interface
}
