% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasview-package.R
\docType{package}
\name{atlasview-package}
\alias{atlasview}
\alias{atlasview-package}
\title{atlasview: View the Co-Occurring Disease Results from the Disease Atlas}
\description{
The purpose of this app is to get the views of clinicians on the co-occurring disease results in the Disease Atlas.
}
\author{
\strong{Maintainer}: Asif Tamuri \email{a.tamur@ucl.ac.uk}

Authors:
\itemize{
  \item Milan Malfait \email{m.malfait@ucl.ac.uk}
}

}
\keyword{internal}
