% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlasview_mock_data}
\alias{atlasview_mock_data}
\title{Atlasview mock data}
\format{
\subsection{\code{atlasview_mock_data}}{

A List with the following components:
\describe{
\item{specialties}{data.frame of of disease specialties with their associated code and colour}
\item{specialty_colours}{named character vector of specialty colours}
\item{MM_res}{data.frame of randomised co-occurring diseases data}
\item{n_dis_spe}{data.frame of randomised patient count data}
\item{index_diseases}{data.frame of index diseases metadata}
}
}
}
\usage{
atlasview_mock_data
}
\description{
Randomised mock data for use in tests and examples.
}
\keyword{datasets}
