% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_plot}
\alias{circos_plot}
\title{Make circos plot for a given index disease}
\usage{
circos_plot(atlasview_data, selected_index_disease, svg_filepath = NULL)
}
\arguments{
\item{atlasview_data}{list of atlasview data, typically from \code{\link[=get_atlasview_data]{get_atlasview_data()}}}

\item{selected_index_disease}{character string, the selected index disease as code}

\item{svg_filepath}{optional filepath. If provided, the plot will be written to the path.}
}
\description{
Make circos plot for a given index disease
}
