% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_caterpillar.R
\name{create_specialty_filter}
\alias{create_specialty_filter}
\title{Generate options for the caterpillar specialty filter}
\usage{
create_specialty_filter(cooccurring_diseases)
}
\arguments{
\item{cooccurring_diseases}{data.frame, typically from \code{\link[=get_cooccurring_diseases]{get_cooccurring_diseases()}}.}
}
\value{
Character vector of specialties, to use in the \code{"filter"} input box for the Caterpillar
tab
}
\description{
Generate options for the caterpillar specialty filter
}
\keyword{internal}
