% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_createProjectModal.R
\name{createProjectModal}
\alias{createProjectModal}
\title{createProjectModal UI function
@description
Once logged in, a manager can choose to open the createProjectModal to generate a new project in the database.
The UI contains a form with creating a new project. Once that form has been filled out, they can add layers associated with the solution
and the solution layer.}
\usage{
createProjectModal(id)
}
\value{
A shiny modal dialog UI
}
\description{
createProjectModal UI function
@description
Once logged in, a manager can choose to open the createProjectModal to generate a new project in the database.
The UI contains a form with creating a new project. Once that form has been filled out, they can add layers associated with the solution
and the solution layer.
}
