% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_projects_database.R
\name{find_projects_database}
\alias{find_projects_database}
\title{Find projects}
\usage{
find_projects_database(x, user_groups = "public")
}
\arguments{
\item{x}{\code{character} directory path.}

\item{user_groups}{\code{character} vector of project group names available that
can be imported.
Defaults to \code{"public"}.}
}
\description{
Find project files in a directory.
}
\details{
query projects available based on user type
}
\examples{
# find directory with built-in projects
d <- system.file("extdata", "projects", package = "wheretowork")

# list projects in directory
find_projects(d)
}
