% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_specialties.R
\name{generate_pagetitle}
\alias{generate_pagetitle}
\title{Generate app page title}
\usage{
generate_pagetitle(atlasview_data, selected_specialty, selected_index_disease)
}
\arguments{
\item{atlasview_data}{data.frame containing atlasview data, typically from \code{\link[=get_atlasview_data]{get_atlasview_data()}}}

\item{selected_specialty}{character, the selected specialty}

\item{selected_index_disease}{character, the selected index disease}
}
\value{
Character string of the format \code{"AtlasViews: {specialty} \\u2192 {disease}"}
}
\description{
Generate the app's page title using the selected specialty and index disease
}
\keyword{internal}
