% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_atlasview_data}
\alias{get_atlasview_data}
\title{Load, process, and return data for specialties and diseases}
\usage{
get_atlasview_data()
}
\description{
Load, process, and return data for specialties and diseases
}
