% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_cooccurring_diseases}
\alias{get_cooccurring_diseases}
\title{Get the co-occurring diseases}
\usage{
get_cooccurring_diseases(
  all_diseases,
  index_disease,
  specialty = NULL,
  specialty_filter = NULL
)
}
\arguments{
\item{all_diseases}{data.frame containing the the full disease dataset, typically from
\code{get_atlasview_data()$MM_res}}

\item{index_disease}{character, the code of the selected input disease}

\item{specialty}{optional character, the specialty code on which to focus}

\item{specialty_filter}{optional character vector, when provided, only co-occurring diseases from
within these specialties will be shown}
}
\value{
A subset of the input data for the co-occurring diseases.
}
\description{
Return the co-occurring diseases for a given index disease from the full data set.
Used to generate caterpillar and circos plots.
}
\keyword{internal}
