% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_filepath}
\alias{get_data_filepath}
\title{Get the full path to a AtlasView data file}
\usage{
get_data_filepath(filename)
}
\arguments{
\item{filename}{name of a file}
}
\description{
Get the full path to a AtlasView data file
}
