% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_specialties.R
\name{get_index_diseases}
\alias{get_index_diseases}
\title{Get available index diseases for given specialty}
\usage{
get_index_diseases(index_diseases, specialty)
}
\arguments{
\item{index_diseases}{data.frame of index diseases}

\item{specialty}{selected specialty code}
}
\description{
Get available index diseases for given specialty
}
\keyword{internal}
