% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Fetch Weather for a City in Sweden}
\usage{
get_weather(city_name)
}
\arguments{
\item{city_name}{Name of the Swedish city}
}
\value{
A list containing the current weather data (temperature, windspeed, etc.) for the specified city.

A list with the current weather information for the city
}
\description{
Fetch Weather for a City in Sweden

Get Weather for a Swedish City
}
\examples{
get_weather("Stockholm")
}
