% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_url.R
\name{parse_url}
\alias{parse_url}
\title{Extract Specialty and Index Disease from URL Query}
\usage{
parse_url(url)
}
\arguments{
\item{url}{The query string. It can have a leading \code{"?"} or not}
}
\description{
Extract Specialty and Index Disease from URL Query
}
\keyword{internal}
