% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_project.R
\name{projectUI}
\alias{projectUI}
\title{Project Module}
\usage{
projectUI(id)
}
\arguments{
\item{id}{Module identifier}

\item{client}{GraphQL client}

\item{auth_token}{Authentication token reactive}

\item{user_info}{User information reactive}

\item{projects_data}{Projects data reactive}
}
\description{
UI and server logic for project management
}
