% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_adminPage.R
\docType{data}
\name{server_adminPage}
\alias{server_adminPage}
\title{Admin Page Server}
\format{
An object of class \code{{} of length 4.
}
\usage{
server_adminPage
}
\arguments{
\item{id}{\code{character} identifier.}

\item{client}{GraphQL client object.}

\item{auth_token}{Reactive value containing authentication token.}

\item{user_info}{Reactive value containing user information.}
}
\value{
Server logic for admin page.
}
\description{
Server logic for the admin page
}
\keyword{datasets}
