% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_load_solution_database.R
\docType{data}
\name{server_load_solution_database}
\alias{server_load_solution_database}
\title{Server logic for loading solutions from database}
\format{
An object of class \code{{} of length 2.
}
\usage{
server_load_solution_database
}
\description{
Server logic for loading solutions from database
}
\keyword{datasets}
