% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_login.R
\docType{data}
\name{server_login}
\alias{server_login}
\title{Login users}
\format{
An object of class \code{{} of length 10.
}
\usage{
server_login
}
\value{
Invisible \code{TRUE}.
}
\description{
Authorizes user logins
}
\details{
This object is designed to be used within \link{app_server} function
Within the \link{app_server} function, it should be called like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{eval(server_login)
}\if{html}{\out{</div>}}
}
\keyword{datasets}
