% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_write_project.R
\name{write_project}
\alias{write_project}
\title{Write project}
\usage{
write_project(
  x,
  dataset,
  path,
  name,
  spatial_path,
  attribute_path,
  boundary_path,
  mode = "advanced",
  user_groups = "public",
  author_name = NULL,
  author_email = NULL
)
}
\arguments{
\item{x}{\code{list} of \link{Theme}, \link{Weight}, \link{Include}, and \link{Exclude} objects.}

\item{dataset}{\link{Dataset} object.}

\item{path}{\code{character} file path to save the configuration file.}

\item{name}{\code{character} name for the solution.}

\item{spatial_path}{\code{character} file path for the spatial data.}

\item{attribute_path}{\code{character} file path for the attribute data.}

\item{boundary_path}{\code{character} file path for the attribute data.}

\item{mode}{\code{character} mode for running the application.
Defaults to \code{"advanced"}.}

\item{user_groups}{\code{character} vector of user groups than can
access the dataset.
Defaults to \code{"public"}.}

\item{author_name}{\code{character} name of the project author.
Defaults to \code{NULL} such that no author name is encoded in the
project configuration file. This means that the application
will report default contact details.}

\item{author_email}{\code{character} email address of the project author.
Defaults to \code{NULL} such that no email address is encoded in the
project configuration file. This means that the application
will report default contact details.}
}
\value{
Invisible \code{TRUE} indicating success.
}
\description{
Save a project to disk.
}
\examples{
 # find data file paths
 f1 <- system.file(
   "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
   package = "wheretowork"
 )
 f2 <- system.file(
   "extdata", "projects", "sim_raster", "sim_raster_attribute.csv.gz",
   package = "wheretowork"
 )
 f3 <- system.file(
   "extdata", "projects", "sim_raster", "sim_raster_boundary.csv.gz",
   package = "wheretowork"
 )

 # create new dataset
 d <- new_dataset(f1, f2, f3)

 # simulate themes and weights
 th <- simulate_themes(d, 1, 1, 2)
 w <- simulate_weights(d, 1)

 # combine themes and weights into a list
 l <- append(th, w)

 # save project
 write_project(
   x = l,
   name = "example",
   dataset = d,
   path = tempfile(),
   spatial_path = tempfile(fileext = ".tif"),
   attribute_path = tempfile(fileext = ".csv.gz"),
   boundary_path = tempfile(fileext = ".csv.gz")
 )
}
